split spatial index test classes into separate files
